Cleaned up all the code :D
