+++ packages/ombits-client/src/interfaces/action.ts
@@ -1,0 +1,4 @@
+export interface Action<PayloadType> {
+  type: string;
+  payload: PayloadType;
+}
