+++ packages/ombits-client/src/pages/planning/action-types.tsx
@@ -1,0 +1,4 @@
+export const ADD_NEW_BIT_BELOW = 'PLANNING_ADD_NEW_BIT_BELOW';
+export const INDENT_BIT = 'PLANNING_INDENT_BIT';
+export const UNINDENT_BIT = 'PLANNING_UNINDENT_BIT';
+export const UPDATE_BIT_TITLE = 'PLANNING_UPDATE_BIT_TITLE';
