+++ packages/ombits-client/src/pages/planning/index.ts
@@ -1,0 +1,1 @@
+export * from './planning';
