--- packages/ombits-client/src/utils/search-bits.ts
@@ -1,20 +1,0 @@
-import { Bit } from '../interfaces/bits';
-
-// Not used but keeping for ref
-export const searchBits = (searchArray: Bit[], searchId: string) => {
-  // let res = null;
-  // searchArray.forEach((obj: Bit) => {
-  //   // If we found the key we're looking for, return the matching value.
-  //   if (obj.id === searchId) {
-  //     res = obj;
-  //     return obj;
-  //   // If we've got a nested object, recursively call lookup on it.
-  //   } else if (obj.children.length > 0) {
-  //     const result = searchBits(obj.children, searchId);
-  //     if (result) {
-  //       res = result;
-  //     }
-  //   }
-  // });
-  // return res;
-};
